Give plugins a few seconds to re-register after kubelet restart
.. if the checkpoint file indicates there should be running
device plugins around.

This could have been a "hot" kubelet restart that will check
all running pods have their devices ready and evict them if
that is not the case.

However the device plugins need a bit of a time to properly
re-register after losing the connection to kubelet.

The behavior is a side effect of a more stringent device health checking from
https://github.com/kubernetes/kubernetes/pull/116376